Aplicacion/requirements.txt
mysql-connector-python==8.0.28
python-dotenv==0.20.0
schedule==1.1.0
elasticsearch
pandas
schedule